+++ CHANGELOG.md
@@ -11,10 +11,12 @@ Most recent change on the bottom.
 ### Added
 - `NequIPCalculator` now handles per-atom energies
 - Added `initial_model_state_strict` YAML option
+- `load_model_state` builder
 - fusion strategy support
 
 ### Changed
 - Disallow PyTorch 1.9, which has some JIT bugs.
+- `nequip-deploy build` now requires `--train-dir` option when specifying the training session
 
 ### Fixed
 - Better error in `Dataset.statistics` when field is missing